rho-31 polished code fixed issue with default deployment properties added into deployment of uploaded CSV
