+++ frontend/src/pages/operatorBundlePage/OperatorBundlePage.js
@@ -3,8 +3,7 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import { connect } from 'react-redux';
 import * as _ from 'lodash-es';
-import JSZip from 'jszip';
-import { safeDump } from 'js-yaml';
+
 import { helpers } from '../../common/helpers';
 import { reduxConstants } from '../../redux/index';
 import EditorSection from '../../components/editor/EditorSection';
@@ -12,14 +11,9 @@ import ManifestUploader from '../../components/editor/manfiestUploader/ManifestU
 import { operatorFieldDescriptions, operatorObjectDescriptions } from '../../utils/operatorDescriptors';
 import OperatorEditorSubPage from './OperatorEditorSubPage';
 import PreviewOperatorModal from '../../components/modals/PreviewOperatorModal';
-import {
-  EDITOR_STATUS,
-  operatorNameFromOperator,
-  yamlFromOperator,
-  getMissingCrdUploads,
-  filterValidCrdUploads
-} from './bundlePageUtils';
-import { defaultOperator, validateOperator, removeEmptyOptionalValuesFromOperator } from '../../utils/operatorUtils';
+import { EDITOR_STATUS } from './bundlePageUtils';
+import { defaultOperator, validateOperator } from '../../utils/operatorUtils';
+import OperatorBundleDownloader from '../../components/editor/BundleDownloader';
 
 class OperatorBundlePage extends React.Component {
   state = {
@@ -54,93 +48,6 @@ class OperatorBundlePage extends React.Component {
     this.props.history.push('/bundle/yaml');
   };
 
-  createPackageFile = (operatorPackage, name) => {
-    let packageYaml = '';
-
-    try {
-      packageYaml = safeDump({
-        packageName: operatorPackage.name,
-        channels: [
-          {
-            name: operatorPackage.channel,
-            currentCSV: name
-          }
-        ]
-      });
-    } catch (e) {
-      console.error("Can't create package file.");
-    }
-    return packageYaml;
-  };
-
-  scrollToUploader = () => {
-    const uploader = document.getElementById('oh-operator--editor-page__manifest-uploader');
-
-    // scroll to uploader if it exists
-    if (uploader) {
-      uploader.scrollIntoView();
-    } else {
-      // fallback to top in case something change
-      window.scroll({ top: 0 });
-    }
-  };
-
-  generateCSV = () => {
-    const { operator, uploads, operatorPackage } = this.props;
-
-    const hasMissingCrdUploads = getMissingCrdUploads(uploads, operator).length > 0;
-
-    // do not allow dowloading bundle with missing CRDs
-    if (hasMissingCrdUploads) {
-      this.scrollToUploader();
-      return;
-    }
-
-    // remove values which are part of default operator, but are invalid
-    const cleanedOperator = removeEmptyOptionalValuesFromOperator(operator);
-
-    let operatorYaml;
-    try {
-      operatorYaml = yamlFromOperator(cleanedOperator);
-    } catch (e) {
-      operatorYaml = '';
-    }
-    const name = operatorNameFromOperator(cleanedOperator);
-    const shortName = _.get(operator, 'metadata.name');
-
-    const uploadedCrds = filterValidCrdUploads(uploads);
-    const packageYaml = this.createPackageFile(operatorPackage, name);
-
-    const zip = new JSZip();
-    zip.file(`${name}/${shortName}.package.yaml`, packageYaml);
-    zip.file(`${name}/${name}.clusterserviceversion.yaml`, operatorYaml);
-
-    uploadedCrds.forEach(crd => {
-      let crdYaml = '';
-      let crdName = '';
-
-      try {
-        crdYaml = safeDump(crd.data);
-        crdName = crd.data.metadata.name;
-      } catch (e) {
-        console.warn(`Can't convert crd to yaml for ${crdName}`);
-      }
-
-      zip.file(`${name}/${crdName}.crd.yaml`, crdYaml);
-    });
-
-    zip.generateAsync({ type: 'base64' }).then(
-      base64 => {
-        this.generateAction.href = `data:application/zip;base64,${base64}`;
-        this.generateAction.download = `${_.get(cleanedOperator, 'spec.displayName')}.bundle.zip`;
-        this.generateAction.click();
-      },
-      err => {
-        console.error(err);
-      }
-    );
-  };
-
   hidePreviewOperator = () => {
     this.setState({ previewShown: false });
   };
@@ -163,10 +70,6 @@ class OperatorBundlePage extends React.Component {
     showConfirmModal(this.doClearContents);
   };
 
-  setGenerateAction = ref => {
-    this.generateAction = ref;
-  };
-
   renderHeader = () => (
     <React.Fragment>
       <div className="oh-operator-editor-page__header">
@@ -181,21 +84,23 @@ class OperatorBundlePage extends React.Component {
   );
 
   renderButtonBar() {
-    const { operator } = this.props;
+    const { operator, uploads, operatorPackage } = this.props;
     const { validCSV, sectionsValid } = this.state;
 
     const isDefault = _.isEqual(operator, defaultOperator);
     const okToDownload = validCSV && sectionsValid;
 
-    const downloadClasses = classNames('oh-button oh-button-primary', { disabled: !okToDownload });
     const clearClasses = classNames('oh-button oh-button-secondary', { disabled: isDefault });
 
     return (
       <div className="oh-operator-editor-page__button-bar">
         <div>
-          <button className={downloadClasses} disabled={!okToDownload} onClick={this.generateCSV}>
-            Download Operator Bundle
-          </button>
+          <OperatorBundleDownloader
+            operator={operator}
+            uploads={uploads}
+            operatorPackage={operatorPackage}
+            disabled={!okToDownload}
+          />
           <button className="oh-button oh-button-secondary" onClick={this.onEditCSVYaml}>
             Edit CSV in YAML
           </button>
@@ -315,7 +220,6 @@ class OperatorBundlePage extends React.Component {
         {this.renderOperatorInstallation()}
         {this.renderPackageInfo()}
         <PreviewOperatorModal show={previewShown} yamlOperator={operator} onClose={this.hidePreviewOperator} />
-        <a className="oh-operator-editor-page__download-link" ref={this.setGenerateAction} />
       </OperatorEditorSubPage>
     );
   }
